fix: macOS/iOS build support - Metal RHI, CMake, and platform compatibility
- CMakeLists.txt: make APPLE exclusion regex case-insensitive to match
  actual lowercase directory names on APFS (interface, vulkan, dx12, passes)
- WindowSystem: add Apple platform stub returning nullptr, guard null
  pointer in Initialize (Cocoa/UIKit host manages the window)
- MetalRHI: add missing destructor/virtual function implementations,
  replace unique_ptr with raw pointer to avoid ObjC type exposure
- RHIFactory: rename .cpp -> .mm for ObjC++ compilation required by
  Metal framework headers
- RenderType/RenderObject: add missing <EASTL/vector.h> includes
- RenderGpuResources: define VmaAllocation as void* for non-Vulkan platforms
- UIPass: remove unnecessary static_cast<int32_t> to fix narrowing warning
- macos_build_zruntime_shared.sh: minor updates

+++ engine/Source/Runtime/Function/Render/Passes/UIPass.cpp
@@ -423,8 +423,8 @@ void UIPass::Draw()
                           1.0f};
     RHIRect2D scissor {0,
                        0,
-                       static_cast<int32_t>(m_Rhi->GetSwapchainInfo().extent.width),
-                       static_cast<int32_t>(m_Rhi->GetSwapchainInfo().extent.height)};
+                       m_Rhi->GetSwapchainInfo().extent.width,
+                       m_Rhi->GetSwapchainInfo().extent.height};
 
     m_Rhi->CmdBindPipelinePFN(command_buffer, RHI_PIPELINE_BIND_POINT_GRAPHICS, m_RenderPipelines[0].pipeline);
     m_Rhi->CmdSetViewportPFN(command_buffer, 0, 1, &viewport);
